==> codes/00_independent_codes.ipynb <==
### Unemployment Rate 2021-2024

# Unemployment Rate (city-level)
import pandas as pd

df_work1 = pd.read_csv("data/raw/data_base/unemployment-rates-24.csv")
df_work2 = pd.read_csv("data/raw/data_base/unemployment-rates-23.csv")
df_work3 = pd.read_csv("data/raw/data_base/unemployment-rates-22.csv")
df_work4 = pd.read_csv("data/raw/data_base/unemployment-rates-21.csv")
df_work5 = pd.read_csv("data/raw/data_base/unemployment-rates-19.csv")
df_work6 = pd.read_csv("data/raw/data_base/unemployment-rates-18.csv")

df_work1.head()

df_work24 = df_work1[["NAME","S2301_C04_001E"]].iloc[1:].rename(columns={"S2301_C04_001E": "2024"})
df_work23 = df_work2[["NAME", "S2301_C04_001E"]].iloc[1:].rename(columns={"S2301_C04_001E": "2023"})
df_work22 = (df_work3[["NAME", "S2301_C04_001E"]].iloc[1:].rename(columns={"S2301_C04_001E": "2022"}))
df_work21 = (df_work4[["NAME", "S2301_C04_001E"]].iloc[1:].rename(columns={"S2301_C04_001E": "2021"}))
df_work19 = (df_work5[["NAME", "S2301_C04_001E"]].iloc[1:].rename(columns={"S2301_C04_001E": "2019"}))
df_work18 = (df_work6[["NAME", "S2301_C04_001E"]].iloc[1:].rename(columns={"S2301_C04_001E": "2018"}))

df_work19.head()

df_work = (
    df_work24
    .merge(df_work23, on="NAME", how="left")
    .merge(df_work22, on="NAME", how="left")
    .merge(df_work21, on="NAME", how="left")
    .merge(df_work19, on="NAME", how="left")
    .merge(df_work18, on="NAME", how="left")
)

df_work.head()

df_work_clean = df_work.melt(
    id_vars="NAME",                               
    value_vars=["2018", "2019",
                "2021", "2022",
                "2023", "2024"],    
    var_name="year",                              
    value_name="unemployment rate" 
)
df_work_clean.head()

df_work_clean = df_work_clean.rename(columns={"NAME": "city"})
df_work_clean.head()

df_work_states = df_work_clean.copy()
df_work_states[["city","state"]] = df_work_states["city"].str.rsplit(", ", n=1, expand=True)
df_work_states.head()

df_work_states.shape

### Median Household Income 2021-2024

# Median Household Income

df_y1 = pd.read_csv("data/raw/data_base/median_income2023.csv")
df_y2 = pd.read_csv("data/raw/data_base/median_income2024.csv")
df_y3 = pd.read_csv("data/raw/data_base/median_income2021.csv")
df_y4 = pd.read_csv("data/raw/data_base/median_income2022.csv")
df_y5 = pd.read_csv("data/raw/data_base/median_income2018.csv")
df_y6 = pd.read_csv("data/raw/data_base/median_income2019.csv")

df_y6.head()

df_y1_clean = df_y1[["NAME", "B19013_001E"]].iloc[1:].rename(columns={"B19013_001E": "income_2023"})
df_y2_clean = df_y2[["NAME", "B19013_001E"]].iloc[1:].rename(columns={"B19013_001E": "income_2024"})
df_y3_clean = df_y3[["NAME", "B19013_001E"]].iloc[1:].rename(columns={"B19013_001E": "income_2021"})
df_y4_clean = df_y4[["NAME", "B19013_001E"]].iloc[1:].rename(columns={"B19013_001E": "income_2022"})
df_y5_clean = df_y5[["NAME", "B19013_001E"]].iloc[1:].rename(columns={"B19013_001E": "income_2018"})
df_y6_clean = df_y6[["NAME", "B19013_001E"]].iloc[1:].rename(columns={"B19013_001E": "income_2019"})

df_y6_clean.head()

income_merged = (
    df_y1_clean
    .merge(df_y2_clean, on="NAME")
    .merge(df_y3_clean, on="NAME")
    .merge(df_y4_clean, on="NAME")
    .merge(df_y5_clean, on="NAME")
    .merge(df_y6_clean, on="NAME")
)

income_merged.head()

income_merged = income_merged.rename(columns={"NAME": "city"})
income_merged = income_merged[
    ["city", "income_2018", "income_2019",
     "income_2021", "income_2022", "income_2023", "income_2024"]
]
income_merged.head()

# change the table to only three column
income_clean = income_merged.melt(
    id_vars="city",
    value_vars=["income_2018", "income_2019",
                "income_2021", "income_2022",
                "income_2023", "income_2024"],
    var_name="year",
    value_name="income"
)

income_clean.head()

income_clean["year"] = income_clean["year"].str[-4:].astype(int)
income_clean.head()

# add state columns
income_clean[["city", "state"]] = income_clean["city"].str.rsplit(", ", n=1, expand=True)
income_clean.head(10)

income_clean.shape

df_work_states.shape

# Merge the two datasets

income_clean["year"] = income_clean["year"].astype(int)
df_work_states["year"] = df_work_states["year"].astype(int)

merged_panel = income_clean.merge(
    df_work_states,
    on=["city", "year", "state"],   # ensure exact match
    how="inner"
)

merged_panel.head()

merged_panel.shape

df_age_sex_edu = pd.read_csv("data/raw/data_base/age_sex_edu_df.csv")
df_age_sex_edu.head()

merged_final = df_age_sex_edu.merge(
    merged_panel,
    on=["city", "year", "state"],   # ensure exact match
    how="inner"
)

merged_final.head()

merged_final.shape

merged_panel.to_csv("data/cleaned/independent_variables.csv", index=False)
